Error trying to simulate with visualization export
Trying to simulate cell division with the following
tutorial https://biodynamo.github.io/user/visualization/
succeeded when the bdm.toml was not included. However,
when the bdm.toml file is created according to the tutorial
the error displayed in "ERROR.txt" is shown. It looks to be
a bad_alloc within paraview itself. It may have been an issue
with an outdated paraview, so I proceeded to update
and rerun. Unfortunately the erroneous output persists.